Initial question timer app
We can add the questions list to the app, the time will count when we start, we can move to the next question before timeout.
The timer will be displayed in 2 place, the first timer is the total time of all question we need to answer, the second is the limit time of each question.
Submit form when we finish answer the question and the list of answer will be showed.