Add multiple choices & change list data file
Add render function for multiple choices. Now we can add 2 options of question type that are write the answer or choose the given answer.

We can change the type of question by pass the 'list' property in init function.

Move the data questions to data.js because the file index.js is a module, it should only contain code related to that module. Separating code and data is easier for people working on the same file, and also for people who will later work on it. If we not seperated then when the module has grown larger, and data can be added more, it will be difficult for developers working with this file.